deploy/src/requirements.txt
troposphere>=2.6.3
pyyaml>=5.3.1
